create separate folders for studio and consumer views
