+++ src/views/consumer/ChannelView/ChannelView.tsx
