+++ src/views/consumer/ChannelView/index.ts
