+++ src/views/consumer/index.ts
@@ -1,0 +1,8 @@
+import HomeView from './HomeView'
+import VideoView from './VideoView'
+import ChannelView from './ChannelView'
+import VideosView from './VideosView'
+import ChannelsView from './ChannelsView'
+import SearchOverlayView from './SearchOverlayView'
+
+export { HomeView, VideoView, SearchOverlayView, ChannelView, VideosView, ChannelsView }
