+++ src/views/studio/index.ts
@@ -1,2 +1,3 @@
 import StudioView from './StudioView'
+
 export { StudioView }
